Enable clone -l on Windows by providing a hard link function.
NOTE: 
This changes what link() does on Windows from doing
a copy to doing a hard link or failing. This is OK because
it's what linux does: the fail is detected and a copy done.
Also, link is not used for locking on Windows as it is on Linux
so shifting the link() command does not affect locking on Windows.

NOTE: t.timestamps was not uncommented because Rick has an RTI with a better
fix and I want it to automerge.

bk: 45832868R4TjbishZhvIu8n3Q9Sw-w