NESTED: add module support to clone.
This is a first pass and will likely evolve during review.
The invariant is that BitKeeper/log/MODULES has to match
what is in the repo or the repo has to be fully populated
(and the file should not exist or get removed by check).
We're still moving towards that.

We punted on the MODULES file changing in your parent and
having that change noticed by a pull.  It would be nice if
we had a bk changes -R -M or something that listed the 
differences between modules files.

bk: 4810a5ffFf7EYp-2ue_ZoGACeaHBSQ