NESTED: make bk -N command work in nested like bk -r command.
This cset has some questionable choices in it but the positives
outweigh the negatives and I think we can fix the negatives in
the future.  An example of where things don't work would be
an export -tpatch where the end points include file[s] that
belong to a component that moved; in that case the historic
component prefix would not be correct, it will show the current
prefix.  

LIBC: fix getopt so it can do -? for bk.c

bk: 4a0895b5Qk-_R_VF-llj-q1Cm1fMLg