blobfs: fix t.park and some diff/patch issues.
My remapping stuff didn't take patch/diffutils into account.  Wayne may
want to take a different approach.

unpark had a problem with buffers that I don't completely understand, 
this does not seem to be blobfs related and it's in an area that is 
unchanged and a test that is unchanged.  I "fixed" fullname and stuck
an assert in concat_path.

bk: 4a1ca83d21k9l46zIswviYhs7pDj5w