triggers: implement inheritence and bugfixes.
$PRODUCT means look in the product.
$REPO_TYPE set in triggers.

Various fixes to make the triggers work better with bk-locks.
(lm, tests from rob)
---
REVIEW - left t.ensemble-triggers in a failing state.

I added a pre-resolve trigger and it showed that the product
ran the trigger in the RESYNC dir and the components did not.

In trigger.c, there are a number of lines which special case
the resolve trigger in the inner loop of running through
triggers:

	if (streq(what, "resolve") && streq(dirs[i], ".")) {
		// .. stuff like chdir(ROOT2RESYNC);

This only works for trigger path of '.' and for all others,
the triggers are run in the repo.  I tried quickly moving
the cd out of the loop and kept hitting things that would
break.  Before digging into it more, I decided to push this
and keep going with the breadth of testing and if no one else
fixes it, come back to it later.

Minor note: Added push and pull which involve cloning components.
Notice that in the file url, the push shows the 'backwardness'
that lm mentions elsewhere in the file: the incoming and outgoing
for the component do not match the client.  In the bk:// case,
they do match (because it is using the rclone mode).

REVIEW part 2 - re-arrange the beginning of triggers to be able
to cd to RESYNC before getting proj root and calling to process
the config file entry for triggers.  This lets the cd to and
from the RESYNC dir happen parenthetically around all the processing
and make the special case less special of a case.  The alternative
is to track down all the callers and do the cd there.  Is it just
push and pull?  Probably a cleaner way to go, but for now, passing
regressions again.
(rick)
---
REGRESSION fix: redhat71 fails some of the time because of a race
in a pull with a component clone between the post-outgoing of the
remote product and the pre-outgoing of the remote clone trigger.
I'm not advocating this fix, but will sleep on it and something
better will surface or others will have suggestions.
(rick)

bk: 4ae31652aHP5EUHDStpIs5FGOB_s8Q