When we are pulling from a non-gate (safe mode) and find components that exist in remote, don't exist in local, and won't be populated by the pull, we used to fail the pull and ask the user to populate them to prevent work from being lost.
This cset makes it so rather than failing the pull, we look for those
components in gates (using urllist) and if we can find them in a gate
we allow the pull to proceed without populating those components.

= IMPLEMENTATION =

All I (ob) really wanted to do was:

{{{
@@ -692,9 +710,21 @@ pull_ensemble(remote *r, char **rmt_alia
 
 	if ((opts.safe == 1) || ((opts.safe == -1) && !getenv("BKD_GATE"))) {
 		char	**missing = 0;
+		int	flags = URLLIST_GATEONLY;
+
+		if (opts.quiet) flags |= SILENT;
 
 		EACH_STRUCT(n->comps, c, j) {
-			if (!c->present && !c->alias && c->remotePresent) {
+			/*
+			 * !c->alias           = I'm not going have it
+			 * c->remotePresent    = Remote has it
+			 * (opts.safe != 1 &&  = user explicitly asked for
+			 *                       --safe
+			 * || !urllist_find()) = we can't find it in a gate
+			 */
+			if (!c->alias && c->remotePresent &&
+			    ((opts.safe == 1) ||
+				!urllist_find(n, c, flags, 0))) {
 				char	**alist = c->data;
 
 				EACH(alist) {
}}}

But the plumbing wasn't there, so it took Wayne, Rick, and myself 
a while to add a programatic API to bk remote (remote_cmd()), do
a new an improved urllist/urlinfo API, and plug everything back
together so that the fore-metioned diff could work. But that diff
is really all this work is trying to accomplish.

bk: 4d5d55f1hTPJpxePAIK58uzmzthRog