Add proj_tipkey() and friends to find the current tip key of the ChangeSet file.  Checks the file time/size to make sure the cached data is accurate.
Use this tipkey in the nested_init() cache instead of relying
on the ChangeSet file size/mtime.

This is partly a fix needed for the order1 tree where size/mtime
of the ChangeSet file can't be used to determine if the ChangeSet
file has changed or now.  Now we add code to make sure TIP is
maintained and then use the tipkey in caches.

bk: 4ee0e8d9_c3KEi4WCFN-xbVy40CWmA