I've seen a few of these recently:
 @@ -417,8 +417,10 @@ exclude(char *cmd_prefix, int verbose)
 findcmd(int ac, char **av)
 {
 	int	i;
 +	char	*av0;

 	if (ac == 0) return (-1);
 +	av0 = av[0];
 	for (i = 0; cmds[i].name; ++i) {
 		if (strcasecmp(av[0], cmds[i].name) == 0) {
 			if (streq(av[0], "pull")) av[0] = "remote pull";

 Notice how the change is attributed to the exclude() function
 and not to the findcmd() function. Well, this fixes that.
---
Add a way to pass the number of lines of context
to ndiff and weave it through the API.
---
Pass the line number of the first diff to the printHeader()
function rather than the line number of the first line 
we're printing. The idea is that when the printHeader() function
looks for function names (for diff -p) it will match functions
that appear in the diff context.

bk: 50770bd1hxQFXW8vw2rhuOK3_6-tCw