Add code to check to get a repository write lock before it starts writing sfiles for autofix, cset marking, or heap repacking.
When check is called internally we normally already have a write lock
so we add BK_NO_REPO_LOCK=YES when calling internally.  Then check
was extended so when BK_NO_REPO_LOCK is used we check to see if we
_really_ have a write lock.

The majority of the cset is fixing our locking code so the assertion
above passes and we are really using locking as expected.

bk: 51351a04cb6zRbZV29fIL8FvFMgTuQ