Fix for a deadlock.
When Windows Explorer is showing a repository, it will sit on
directories like PENDING and RESYNC as they appear which will
then interfere with BK operations and in some cases will leave
behind an empty RESYNC directory (which will outright prevent
various BK operations).

Early attempts to fix this problem focused on treating PENDING
and RESYNC as special cases but we can also see symptoms of the
same problem when undo/unpull cleans up empty directories.

The root problem is that in our Windows implementation of rmdir
will attempt to ensure that nobody has a directory open before
removing it.  This is done to ensure POSIX semantics.

Also, while here, reduce the number of retries.  It was excessive.

Experiments have shown that Explorer can hold directories open
such that our attempts to grab a lock fail.  So now we unconditionally
attempt the RemoveDirectory() call and consider retrying when/if
that fails (as it will if somebody is really holding the directory).

While there (wapi_intf.c), reduce the number of retries.  It
was excessive.  And fix a couple of compiler warnings.

Added bk _rmdir command to allow for manual testing.

Add a t. test to show that the rmdir retry loop still works.

Add an i. test that attempts to show that nt_rmdir will work
even when Explorer is watching.

Thanks to Rick and Rob for reviews and tweaks.

bk: 5386080azPHIeb7VeNEVuY62Ok0-fg