Perf fix adding static hash table to heap on disk.
When adding new metadata to the heap, previously the code needed to scan all
existing data in order to collapse duplicates. Now for data that we deduplicate
we embed a hash table in the heap whenever we repack. This means the code looking
for duplicates only needs to load new data and then only if the value needed
was new.  This hash table is very small since it doesn't need to store the key in the
hash.

For commit another change was added in this cset that is needed to benefit
from this change. The loop that walks the cset hash looking for previous csets that
touch the same files now compares rootkeys by pointer rather than doing a string
compare. This is faster because we are not paging the entire rootkey table into
memory.

This cset contains lots of input from Rick.

bk: 5429c92c-51pTUgdFxgY-KXDvPu-OA