A 'simple' tweak to the BWEAVE storage format in the ChangeSet file results is large speedups for commands that need to extract information from the ChangeSet weave for a subset of the files. This includes commit, takepatch, rset and others.
As a bonus the cset weave was extended to include a per-rootkey
marker to indicate the oldest deltakey for each rootkey.
This allows commands to exist early without reading the
entire weave.

This change involves a large compatibility footprint.
Ascii cset files are now converted to the new format in
memory, so all cset weave operations in the code can assume
it is in the new binary weave format.

bk: 547faea4yAcmAnoQRbrFW-tgXzTTZw