This is the second part of the config work.
It contains the documentation changes, enhancements to "bk config -v",
and "bk config", as well as the restoring of a config file in repositories
created with just "bk setup" with no arguments. The config file is better
documented and a regression has been added to test for undocumented config
variables.

Changes to defaults (since they are hard to see because I added a field):

  - CHECKOUT, and BAM_CHECKOUT added to confvars.h
  - AUTOFIX default changed from "off" to "on"
  - PARALLEL default changed from 0 to "on"

Whole cset was a collaboration between Wayne, Rick, and Oscar.

bk: 54da7660cpvuFVY69eUtVVCahSCJlw